backend/requirements.txt
fastapi==0.115.11
fitz==0.0.1.dev2
httpx==0.28.1
jose==1.0.0
langchain==0.3.20
loguru==0.7.2
mistralai==1.5.1
google-generativeai==0.3.2
networkx==3.4.2
nltk==3.9.1
passlib==1.7.4
Pillow==11.1.0
protobuf==6.30.1
pydantic==2.10.6
qdrant_client==1.13.3
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
